make NumberArray not use MemoryUtil because that's not available on servers, fix underground pockets using 2 different noise seeds for raw generation vs. regular generation, bump version to 3.11.1.

+++ src/main/java/builderb0y/bigglobe/noise/ScriptedGrid1D.java
@@ -60,7 +60,7 @@ public class ScriptedGrid1D extends ScriptedGrid<Grid1D> implements Grid1D {
 		//workaround for the fact that I *really* don't want to deal
 		//with generating bytecode for try-with-resources at runtime.
 		NumberArray.Direct.Manager manager = NumberArray.Direct.Manager.INSTANCES.get();
-		long used = manager.used;
+		int used = manager.used;
 		try {
 			this.delegate.getBulkX(seed, startX, samples);
 		}
